backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
requests==2.28.2
python-dotenv==1.0.0
openai==1.65.1
